Restructure: promote yaru_colors to the root
The yaru_colors package is the beef of this repo whereas the builder and
generator packages are more like implementation details that rarely need
to be touched. Promoting yaru_colors to the root makes the repo look and
feel like a basic Flutter package and shall reduce the barrier for those
who are not fond of managing the classic monorepo structure with Melos.